Workshop Day1/requirements.txt
python-dotenv
langchain-groq
groq
